Made pointer arithmetic more explicit for clarity
The pointer arithmetic used in calculating offsets was brittle. This commit
addresses that by using more and better named variables and doing a lot of
casting to make sure the memory addresses are being calculated how they were
intended. In particular, a main source of bugs in Suggestomatic has stemmed from
adding bytes to unsigned int*'s and having the memory address delta be off by a
factor of sizeof(unsigned int).

This also changes how the 'end' memory offset is calculated for the last array
in the binary file: we can't look at index+1 to find where the next one starts,
so we special case that in and use the filesize. This could also be fixed by
adding one more entry in the index table that acts as a stop marker, which would
obviate the need for special casing the last iteration. This way was quicker and
more local of a change though.

This commit also removes all the references to 'random_offset' that was intended
to be used to pick a random starting point in the inner loop. The value was
being calculated but not used, so it's been removed until the entire feature is
implemented.