Collect all score results and sort to obtain top 50.
The 'goodmatches' heuristic wasn't cutting it, a full run on our production data
yielded only 10% coverage -- many of the smaller sets just weren't getting
enough. This commit switches to calculating all the scores, using qsort(3) to
grab just the best 50 and write those out. This trades a lot of speed for
quality, but 10% coverage is unacceptable.