Bloom vs Standard performance showdown
Trying to determine at what threshold it makes sense to start using a bloom
filter, and how much time it potentially saves. If the savings are marginal, I'm
in favor of dropping bloom filters altogether, because of the added complexity.
If it's worth it, this will only take a little more polish to merge back into
master.