Multiprocess engine implementation
Building a similarity matrix is very parallelizable, and most computers have
more than one core. This commit introduced a rudimentary pattern that forks
worker processes and attaches to the shared memory that the parent process
mmap'ed. Each worker writes to it's own <output>.X file, where X is the internal
worker ID of that process