Fix offset bug
To find the end offset for an array of set members, we were looking up set_id+1
in the index. This only works if set_ids are contiguous, which they are not
guaranteed to be. This patch adds a next_set_id parameter to load_set_array to
make sure we always find the correct end memory offset.