Add generate_performance_data.py
This script will generate sample input CSV membership data that roughly tries to
mimic the scale of Causes' membership data. The numbers as hardcoded yield ~350
MB of membership data.

Note: this is definitely not optimized for runtime performance; it was optimized
for least amount of time spent coding. It's not efficient, but you only need it
to run once. So kick it off and go grab a burger.